requirements.txt
streamlit
langchain-groq
langchain-core
python-dotenv